refactor frontend architecture
