--- frontend/src/components/Survey/SurveyForm.constants.ts
@@ -1,14 +1,0 @@
-export const FOOD_OPTIONS = [
-  'Pizza',
-  'Pasta',
-  'Pap and Wors',
-  'Other',
-] as const;
-
-export const RATING_OPTIONS = [
-  { value: '1', label: 'Strongly Disagree' },
-  { value: '2', label: 'Disagree' },
-  { value: '3', label: 'Neutral' },
-  { value: '4', label: 'Agree' },
-  { value: '5', label: 'Strongly Agree' },
-] as const;
