+++ frontend/src/components/survey/RatingRow.module.css
