+++ frontend/src/constants/appConfig.ts
