+++ frontend/src/constants/queryKeys.ts
@@ -1,0 +1,3 @@
+export const queryKeys = {
+  results: ['results'] as const,
+} as const;
