+++ frontend/src/constants/routes.ts
@@ -1,0 +1,4 @@
+export const ROUTES = {
+  HOME: '/' as const,
+  RESULTS: '/results' as const,
+} as const;
