+++ frontend/src/hooks/useSubmitSurvey.ts
@@ -1,0 +1,15 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { queryKeys } from '@/constants/queryKeys';
+import { submitSurvey } from '@/services/surveyService';
+
+export function useSubmitSurvey() {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: submitSurvey,
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({ queryKey: queryKeys.results });
+    },
+  });
+}
