+++ frontend/src/pages/SurveyPage.tsx
@@ -1,0 +1,5 @@
+import SurveyForm from '@/components/survey/SurveyForm';
+
+export default function SurveyPage() {
+  return <SurveyForm />;
+}
