+++ frontend/tsconfig.app.json
@@ -28,10 +28,18 @@
       "@/components/*": ["src/components/*"],
       "@/services": ["src/services"],
       "@/services/*": ["src/services/*"],
+      "@/constants": ["src/constants"],
+      "@/constants/*": ["src/constants/*"],
+      "@/hooks": ["src/hooks"],
+      "@/hooks/*": ["src/hooks/*"],
+      "@/layouts": ["src/layouts"],
+      "@/layouts/*": ["src/layouts/*"],
+      "@/pages": ["src/pages"],
+      "@/pages/*": ["src/pages/*"],
+      "@/utils": ["src/utils"],
+      "@/utils/*": ["src/utils/*"],
       "@/validation": ["src/validation.ts"],
       "@/validation/*": ["src/validation/*"],
-      "@/config": ["src/config"],
-      "@/config/*": ["src/config/*"],
       "@shared-root": ["../shared"],
       "@shared-root/*": ["../shared/*"]
     },
